Pull out code from fetching from the recorder to a separate service
