+++ catalogue_pipeline/merger/src/main/scala/uk/ac/wellcome/platform/merger/services/MergerWorkerService.scala
@@ -9,9 +9,6 @@ import uk.ac.wellcome.messaging.sqs.SQSStream
 import uk.ac.wellcome.models.matcher.{MatcherResult, WorkIdentifier}
 import uk.ac.wellcome.models.recorder.internal.RecorderWorkEntry
 import uk.ac.wellcome.models.work.internal.{BaseWork, UnidentifiedWork}
-import uk.ac.wellcome.storage.ObjectStore
-import uk.ac.wellcome.storage.dynamo._
-import uk.ac.wellcome.storage.vhs.{EmptyMetadata, VersionedHybridStore}
 import uk.ac.wellcome.json.JsonUtil._
 
 import scala.collection.Set
@@ -20,9 +17,7 @@ import scala.concurrent.{ExecutionContext, Future}
 class MergerWorkerService @Inject()(
   system: ActorSystem,
   sqsStream: SQSStream[NotificationMessage],
-  vhs: VersionedHybridStore[RecorderWorkEntry,
-                            EmptyMetadata,
-                            ObjectStore[RecorderWorkEntry]],
+  playbackService: RecorderPlaybackService,
   merger: Merger,
   messageWriter: MessageWriter[BaseWork]
 )(implicit ec: ExecutionContext)
@@ -52,15 +47,9 @@ class MergerWorkerService @Inject()(
   }
 
   private def getFromVHS(
-    matcherResult: MatcherResult): Future[List[Option[RecorderWorkEntry]]] = {
-    val worksIdentifiers = getWorksIdentifiers(matcherResult)
-    for {
-      maybeWorkEntries <- Future.sequence(worksIdentifiers.toList.map {
-        workId =>
-          getRecorderEntryForIdentifier(workId)
-      })
-    } yield maybeWorkEntries
-  }
+    matcherResult: MatcherResult): Future[List[Option[RecorderWorkEntry]]] =
+    playbackService
+      .fetchAllRecorderWorkEntries(getWorksIdentifiers(matcherResult).toList)
 
   private def sendWorks(mergedWorks: Seq[BaseWork]) = {
     Future
@@ -76,26 +65,6 @@ class MergerWorkerService @Inject()(
       matchedIdentifiers <- matcherResult.works
       workIdentifier <- matchedIdentifiers.identifiers
     } yield workIdentifier
-  }
-
-  private def getRecorderEntryForIdentifier(
-    workIdentifier: WorkIdentifier): Future[Option[RecorderWorkEntry]] = {
-    workIdentifier.version match {
-      case 0 =>
-        Future.successful(None)
-      case _ =>
-        vhs.getRecord(id = workIdentifier.identifier).map {
-          case None =>
-            throw new RuntimeException(
-              s"Work ${workIdentifier.identifier} is not in vhs!")
-          case Some(record) if record.work.version == workIdentifier.version =>
-            Some(record)
-          case Some(record) =>
-            debug(
-              s"VHS version = ${record.work.version}, identifier version = ${workIdentifier.version}, so discarding work")
-            None
-        }
-    }
   }
 
   def stop(): Future[Terminated] = system.terminate()
